+++ average.h
@@ -1,2 +1,2 @@
 void calc_average();
-extern float average[3];
+extern int average[3];
